Modification du controller delete de sauce, supprime bien l'image

+++ controllers/sauce.js
@@ -1,5 +1,5 @@
 const Sauce = require("../models/Sauce");
-
+const fs = require("fs");
 //Fonction POST Création sauce
 exports.createSauce = async (req, res, next) => {
   const sauceNew = await JSON.parse(req.body.sauce);
@@ -58,7 +58,22 @@ exports.modifySauce = (req, res, next) => {
 };
 //Fonction DELETE sauce
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Objet supprimé !" }))
-    .catch((error) => res.status(400).json({ error }));
+  Sauce.findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: "Not authorized" });
+      } else {
+        const filename = sauce.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: "Objet supprimé !" });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ error });
+    });
 };
